+++ ts/packages/json-schema-to-zod/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @composio/json-schema-to-zod
 
+## 0.1.19
+
+### Patch Changes
+
+- b5cc23f: Fix dangerously skip version check in non agentic providers, Throw error instead of process.exit when api key doesn't exist, bump zod-to-json-schema to 3.25.0, which supports "zod/3"
+
 ## 0.1.18
 
 ### Patch Changes